Import live.2004.12.22.tar.gz
- Updated "MPEG4GenericBufferedPacket::nextEnclosedFrameSize() to (i) allow
  for "generic" mode, and (ii) properly check for the absence of an
  "AU Header section".  (Thanks to Erick van Rijk for this suggestion.)
- When receiving a MPEG-2 Transport Stream (either RTP *or* raw-UDP), pass it
  through a "MPEG2TransportStreamFramer", so that "durationInMicroseconds" is
  set appropriately, based on the stream's embedded PCR values.
  (Thanks to Dermot McGahon for this suggestion.)
- Changed the "num_packets_lost" statistics line (printed by "openRTSP -Q")
  to format as an int rather than as an unsigned.  This is in case this value
  is negative (which can happen if duplicate packets are received).
  (Thanks to Norbert Donath for noting this.)

+++ testProgs/playCommon.cpp
@@ -1055,7 +1055,8 @@ void printQOSData(int exitCode) {
 	numPacketsExpected = curQOSRecord->totNumPacketsExpected;
       }
       *env << "num_packets_received\t" << numPacketsReceived << "\n";
-      *env << "num_packets_lost\t" << numPacketsExpected - numPacketsReceived << "\n";
+      *env << "num_packets_lost\t"
+	   << (int)(numPacketsExpected - numPacketsReceived) << "\n";
       
       if (curQOSRecord != NULL) {
 	unsigned secsDiff = curQOSRecord->measurementEndTime.tv_sec
